+++ pages/api/taskCards.ts
@@ -1,9 +1,6 @@
 // apiService.js
 import AXIOS from '@/lib/axios'
-import {
-  TaskCardDataType,
-  TaskCardsPromise,
-} from '@/src/types/dashboard.interface'
+import { TaskCardDataType, TaskCardsPromise } from '@/src/types/dashboard'
 
 /**
  * API 요청에 사용될 공통 인증 헤더 반환
